requirements.txt
agate==1.6.3
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
ariadne==0.23.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
awscli==1.25.78
Babel==2.10.3
backoff==2.2.1
bitarray==2.9.2
blinker==1.8.2
botocore==1.27.77
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
ckzg==2.0.0
click==8.1.7
colorama==0.4.4
confluent-kafka==2.2.0
cycler==0.11.0
cytoolz==0.12.3
databricks-sql-connector==2.0.5
DateTime==4.5
dbt-core==1.2.1
dbt-databricks==1.2.1
dbt-extractor==0.4.1
dbt-spark==1.2.0
distlib==0.3.5
docopt==0.6.2
docutils==0.16
eth-account==0.13.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.8.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.34.4
frozenlist==1.4.1
future==0.18.2
gevent==24.2.1
gql==3.4.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
hexbytes==1.2.1
hologram==0.0.15
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
leather==0.3.4
Logbook==1.5.3
lru-dict==1.3.0
MarkupSafe==2.1.5
mashumaro==2.9
matplotlib==3.5.2
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
multidict==6.0.5
neo4j==5.20.0
networkx==2.8.6
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parsedatetime==2.4
parsimonious==0.10.0
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==5.27.4
psycopg2-binary==2.9.9
pyArango==2.0.1
pyarrow==9.0.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.4.0
PyMySQL==1.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-arango==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.1
pyunormalize==15.1.0
PyYAML==5.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.9.1
rlp==4.0.1
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparams==5.0.0
sqlparse==0.4.2
starlette==0.38.2
strawberry-graphql==0.237.3
supervisor==4.2.5
text-unidecode==1.3
thrift==0.13.0
toml==0.10.2
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.16.3
web3==6.6.1
websockets==13.0
Werkzeug==3.0.4
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.3
